chore: - changed messages page name to fit like other pages - moved messages page to dashboard like others and its navigation
feature:
- better ui for messages page that looks more inline with other pages
